% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{add_timepoints}
\alias{add_timepoints}
\title{Add multiple timepoints from a data frame}
\usage{
add_timepoints(timer, df)
}
\arguments{
\item{timer}{an instance of \link{Timer}}

\item{df}{data frame with following columns:
\itemize{
\item time (\code{numeric}) calendar time
\item arm (\code{character}) unique arm key
\item dropper (\code{integer}) # of subjects to drop
\item enroller (\code{integer}) # of subjects to enroll
}}
}
\description{
Add multiple timepoints from a data frame
}
\examples{
t <- Timer$new(name = "Timer")

timepoints <- data.frame(
time = c(1,2,3.1,4,5,6),
arm = rep("Arm A", 6),
dropper = c(2L, rep(1L, 5)),
enroller = rep(3L, 6)
)

add_timepoints(t, timepoints)
}
