% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trials.R
\name{clone_trial}
\alias{clone_trial}
\title{Clone a Trial Object Multiple Times}
\usage{
clone_trial(trial, n = 1)
}
\arguments{
\item{trial}{\code{Trial} R6 object to clone.}

\item{n}{\code{integer} Number of clones to create. Defaults to \code{1}.}
}
\value{
\code{list} of \code{n} independently cloned \code{Trial} objects.
}
\description{
Creates deep copies of a \code{Trial} R6 object with independent
timer and population instances.
}
\examples{
# Create 5 independent copies of a trial:
# clones <- clone_trial(trial, n = 5)

}
\seealso{
\link{Trial}, \code{\link[=replicate_trial]{replicate_trial()}}, \code{\link[=run_trials]{run_trials()}}.
}
