% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generators.R
\name{gen_plan}
\alias{gen_plan}
\title{Generate plan for trial with allocation, enrollment and dropout}
\usage{
gen_plan(sample_size, arms, allocation, enrollment, dropout)
}
\arguments{
\item{sample_size}{\code{integer} trial sample size}

\item{arms}{\code{character} vector of unique identifier of arms}

\item{allocation}{\code{numeric} vector of arm allocations}

\item{enrollment}{\code{function} PDF of inter-enrollment time}

\item{dropout}{\code{function} PDF of inter-dropout time}
}
\value{
\code{data.frame} timepoints that may be passed to \code{\link[=add_timepoints]{add_timepoints()}}
}
\description{
Generate plan for trial with allocation, enrollment and dropout
}
\examples{
gen_plan(
  sample_size = 100,
  arms = c("A", "B"),
  allocation = c(2,1),
  enrollment = function(n) rexp(n, rate = 0.5),
  dropout = function(n) rexp(n, rate = 0.1)
)

}
