% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trials.R
\name{gen_population}
\alias{gen_population}
\title{Generate a Population Object}
\usage{
gen_population(name, generator, sample_size = 1)
}
\arguments{
\item{name}{\code{character} Population name (e.g., arm label).}

\item{generator}{\code{function} that returns a data.frame-like object.}

\item{sample_size}{\code{integer} Number of subjects to generate.}
}
\value{
\code{Population} R6 object.
}
\description{
Creates a \code{Population} R6 object by executing a generator function.
}
\examples{
gen_control <- function(n) data.frame(id = 1:n, value = rnorm(n), readout_time = 0)
pop <- gen_population("control", gen_control, sample_size = 50)

}
\seealso{
\link{Population}, \code{\link[=replicate_trial]{replicate_trial()}}, \code{\link[=vector_to_dataframe]{vector_to_dataframe()}}.
}
