% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_col_names}
\alias{get_col_names}
\title{Extract Column Names from Populations}
\usage{
get_col_names(populations)
}
\arguments{
\item{populations}{\code{\link{Population}} object or \code{list} of \code{\link{Population}} objects.}
}
\value{
\code{character} vector of unique population column names augmented with
\code{subject_id}, \code{enroll_time}, \code{drop_time}, \code{measurement_time}, and \code{time}.
}
\description{
Collects all data frame column names from one or more populations.
}
\examples{
pop1 <- Population$new(name = "P1", data = data.frame(
id = 1:10,
age = runif(10, 20, 60),
readout_time = 0
))
pop2 <- Population$new(name = "P2", data = data.frame(
id = 1:10,
weight = runif(10, 150, 250),
readout_time = 0
))
get_col_names(list(pop1, pop2))
}
\seealso{
\link{Population}.
}
