% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{prettify_results}
\alias{prettify_results}
\title{Print trial results as a data.frame}
\usage{
prettify_results(results)
}
\arguments{
\item{results}{pass results data field of your \code{Trial}}
}
\description{
Print trial results as a data.frame
}
