% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trials.R
\name{run_trials}
\alias{run_trials}
\title{Run Multiple Trial Objects}
\usage{
run_trials(trials)
}
\arguments{
\item{trials}{\code{list} of \code{Trial} R6 objects.}
}
\value{
\code{list} of results from each trial's \verb{$run()} method.
}
\description{
Executes the \verb{$run()} method for each trial in a list.
}
\examples{
# results <- run_trials(trials)

}
\seealso{
\link{Trial}, \code{\link[=replicate_trial]{replicate_trial()}}.
}
