% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{trigger_by_calendar}
\alias{trigger_by_calendar}
\title{Trigger analysis at a specific calendar time}
\usage{
trigger_by_calendar(cal_time, timer, analysis = NULL)
}
\arguments{
\item{cal_time}{\code{numeric} Calendar time to trigger at}

\item{timer}{\code{Timer} Timer instance to add condition to}

\item{analysis}{\code{function} or \code{NULL} Analysis function to apply}
}
\value{
returns \code{timer} for chaining.
}
\description{
Trigger analysis at a specific calendar time
}
\examples{
t <- Timer$new("timer")
trigger_by_calendar(2, t, analysis = function(df, current_time) nrow(df))

}
