% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{trigger_by_fraction}
\alias{trigger_by_fraction}
\title{Trigger analysis when sample fractionof enrolled subjects is reached}
\usage{
trigger_by_fraction(fraction, timer, sample_size, analysis = NULL)
}
\arguments{
\item{fraction}{\code{numeric} Fraction of target sample (0 < fraction <= 1)}

\item{timer}{\code{Timer} Timer instance to add condition to}

\item{sample_size}{\code{integer} Target sample size}

\item{analysis}{\code{function} or \code{NULL} Analysis function to apply}
}
\value{
returns \code{timer} for chaining.
}
\description{
Trigger analysis when sample fractionof enrolled subjects is reached
}
\examples{
t <- Timer$new("timer")
trigger_by_fraction(0.5, t, sample_size = 100, analysis = function(df, current_time) nrow(df))

}
