% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{vector_to_dataframe}
\alias{vector_to_dataframe}
\title{Convert vector to our dataframe format}
\usage{
vector_to_dataframe(data)
}
\arguments{
\item{data}{\code{numeric} population data in vector format}
}
\value{
\code{data.frame} population data that may be passed to \link{Population}
}
\description{
Convert vector to our dataframe format
}
