% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{add_timepoints}
\alias{add_timepoints}
\title{Add Timepoints to a Timer}
\usage{
add_timepoints(timer, df)
}
\arguments{
\item{timer}{\code{\link{Timer}} instance.}

\item{df}{\code{data.frame} with columns: \code{time} (numeric), \code{arm} (character),
\code{enroller} (integer), \code{dropper} (integer).}
}
\description{
Adds multiple enrollment and dropout events from a data frame.
}
\examples{
t <- Timer$new(name = "Timer")

timepoints <- data.frame(
  time = c(1, 2, 3.1, 4, 5, 6),
  arm = rep("Arm A", 6),
  dropper = c(2L, rep(1L, 5)),
  enroller = rep(3L, 6)
)

add_timepoints(t, timepoints)
}
\seealso{
\link{Timer}, \code{\link[=gen_plan]{gen_plan()}}, \code{\link[=gen_timepoints]{gen_timepoints()}}.
}
