% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{gen_timepoints}
\alias{gen_timepoints}
\title{Generate timepoints for trial with allocation, piece-wise linear enrollment and dropout}
\usage{
gen_timepoints(n, arms, allocation, enrollment, dropout)
}
\arguments{
\item{n}{\code{integer} trial sample size}

\item{arms}{\code{character} vector of unique identifier of arms}

\item{allocation}{\code{numeric} vector of arm allocations}

\item{enrollment}{\code{list} named list with enrollment parameters
\itemize{
\item \code{end_time} vector of end times of duration
\item \code{rate} # of subjects to enroll per unit time within that duration
}}

\item{dropout}{\code{list} named list with dropout parameters
\itemize{
\item \code{end_time} vector of end times of duration
\item \code{rate} # of subjects to enroll per unit time within that duration
}}
}
\value{
\code{data.frame} timepoints that may be passed to \code{\link[=add_timepoints]{add_timepoints()}}
}
\description{
Generate timepoints for trial with allocation, piece-wise linear enrollment and dropout
}
\examples{
gen_timepoints(
  n = 100,
  arms = c("A", "B"),
  allocation = c(2,1),
  enrollment = list(
    end_time = c(4,8,12),
    rate = c(6,12,18)
  ),
  dropout = list(
    end_time = c(5,9,13),
    rate = c(0,3,6)
  )
)
}
