% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_col_names}
\alias{get_col_names}
\title{Get a list of data column names}
\usage{
get_col_names(populations)
}
\arguments{
\item{populations}{\code{list} or \code{Population} Population object(s) to extract column names from}
}
\value{
\code{vector} of column names
}
\description{
Get a list of data column names
}
\examples{
pop1 <- Population$new(name = "P1", data = data.frame(subject_id = 1:10, age = runif(10, 20, 60)))
pop2 <- Population$new(name = "P2", data = data.frame(subject_id = 1:10, weight = runif(10, 150, 250)))
get_col_names(list(pop1, pop2))
}
