% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Population.R
\name{Population}
\alias{Population}
\title{Population: Manage a patient population}
\value{
A new \code{Population} instance.
}
\description{
The \code{Population} class stores subject-level data and manages enrollment and
dropout times for each subject.

Use \code{set_enrolled()} and \code{set_dropped()} to assign enrollment and dropout
times to random subsets of subjects.
}
\examples{
# Basic example: vector input
pop <- Population$new(name = "Control", data = rnorm(10))

pop$n              # number of subjects
head(pop$data)     # generated subject-level data

# Set enrollment for 5 subjects at time = 1
pop$set_enrolled(n = 5, time = 1)

# Drop 2 subjects at time = 3
pop$set_dropped(n = 2, time = 3)

# Reset underlying data
pop$set_data(rnorm(8))


## ------------------------------------------------
## Method `Population$new`
## ------------------------------------------------

Population$new(name = "Intervention", data = rnorm(5))

## ------------------------------------------------
## Method `Population$set_enrolled`
## ------------------------------------------------

pop <- Population$new("Test", rnorm(10))
pop$set_enrolled(n = 4, time = 2)

## ------------------------------------------------
## Method `Population$set_dropped`
## ------------------------------------------------

pop <- Population$new("Test", rnorm(10))
pop$set_enrolled(n = 5, time = 1)
pop$set_dropped(n = 2, time = 3)

## ------------------------------------------------
## Method `Population$set_data`
## ------------------------------------------------

pop <- Population$new("ResetDemo", rnorm(5))
pop$set_data(
  data.frame(
    subject_id = 1:8,
    endpoint = rnorm(8),
    population_name = "ResetDemo"
  )
)
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{name}}{\code{character} Unique identifier for the population.}

\item{\code{data}}{\code{data.frame} Subject-level data frame with columns:
\itemize{
\item \code{subject_id} \code{integer}
\item \code{data} \code{numeric}
\item \code{population_name} \code{character}
\item may contain more columns
}}

\item{\code{enrolled}}{\code{numeric} Vector of enrollment times for each subject.}

\item{\code{dropped}}{\code{numeric} Vector of dropout times for each subject.}

\item{\code{n}}{\code{integer} Number of unique subjects in the population.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Population-new}{\code{Population$new()}}
\item \href{#method-Population-set_enrolled}{\code{Population$set_enrolled()}}
\item \href{#method-Population-set_dropped}{\code{Population$set_dropped()}}
\item \href{#method-Population-set_data}{\code{Population$set_data()}}
\item \href{#method-Population-clone}{\code{Population$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Population-new"></a>}}
\if{latex}{\out{\hypertarget{method-Population-new}{}}}
\subsection{Method \code{new()}}{
Create a new `Population`` instance.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Population$new(name, data = NULL, enrolled = NULL, dropped = NULL, n = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{\code{character} Unique identifier for the population.}

\item{\code{data}}{\code{data.frame} Underlying subject data frame with columns
\code{subject_id}, \code{data}, and \code{population_name}.}

\item{\code{enrolled}}{\code{numeric} enrollment times are always initialized
automatically as \code{NA}.}

\item{\code{dropped}}{\code{numeric} dropout times are always initialized
automatically as \code{NA}.}

\item{\code{n}}{\code{integer} \code{n} is automatically computed from data.}
}
\if{html}{\out{</div>}}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{Population$new(name = "Intervention", data = rnorm(5))
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Population-set_enrolled"></a>}}
\if{latex}{\out{\hypertarget{method-Population-set_enrolled}{}}}
\subsection{Method \code{set_enrolled()}}{
Mark \code{n} enrolled subjects at a given time.

Enrollment applies to subjects who are not yet enrolled \code{NA}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Population$set_enrolled(n, time)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{n}}{\code{integer} Number of subjects to enroll.}

\item{\code{time}}{\code{numeric} Enrollment time.}
}
\if{html}{\out{</div>}}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{pop <- Population$new("Test", rnorm(10))
pop$set_enrolled(n = 4, time = 2)
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Population-set_dropped"></a>}}
\if{latex}{\out{\hypertarget{method-Population-set_dropped}{}}}
\subsection{Method \code{set_dropped()}}{
Mark \code{n} subjects as dropped at a given time.

Dropout applies only to subjects who:
\itemize{
\item have enrolled, and
\item have not already dropped.
}
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Population$set_dropped(n, time)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{n}}{\code{integer} # of subjects to drop}

\item{\code{time}}{\code{numeric} Dropout time}
}
\if{html}{\out{</div>}}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{pop <- Population$new("Test", rnorm(10))
pop$set_enrolled(n = 5, time = 1)
pop$set_dropped(n = 2, time = 3)
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Population-set_data"></a>}}
\if{latex}{\out{\hypertarget{method-Population-set_data}{}}}
\subsection{Method \code{set_data()}}{
Replace the underlying subject data and reset enrollment/dropout status.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Population$set_data(data)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{\code{data.frame} Subject-level data frame with columns:
\itemize{
\item \code{subject_id} \code{integer}
\item \code{data} \code{numeric}
\item \code{population_name} \code{character}
\item may contain more columns
}}
}
\if{html}{\out{</div>}}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{pop <- Population$new("ResetDemo", rnorm(5))
pop$set_data(
  data.frame(
    subject_id = 1:8,
    endpoint = rnorm(8),
    population_name = "ResetDemo"
  )
)
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Population-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Population-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Population$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
