% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxsim-package.R
\docType{package}
\name{rxsim-package}
\alias{rxsim}
\alias{rxsim-package}
\title{rxsim: Reducing friction for Randomized Clinical Trial simulations}
\description{
Simulate phase II and/or phase III clinical trials. It supports various types of endpoints and adaptive strategies.
}
\author{
\strong{Maintainer}: Saumil Shah \email{saumil.shah@boehringer-ingelheim.com}

Authors:
\itemize{
  \item Matthew Valko \email{matthew.valko@boehringer-ingelheim.com}
}

Other contributors:
\itemize{
  \item Ming Yang \email{ming.yang.ext@boehringer-ingelheim.com} [contributor]
  \item Mitchell Thomann \email{mitchell.thomann@boehringer-ingelheim.com} [contributor]
  \item Boehringer Ingelheim Pharma GmbH & Co. KG (\href{https://ror.org/00q32j219}{ROR}) [copyright holder, funder]
}

}
\keyword{internal}
