% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Timer.R
\name{Timer}
\alias{Timer}
\title{Timer: Track timed events across arms}
\description{
A class to collect and query \emph{timepoints} - time-based enrollment and
dropout events - across trial arms.

Use \code{add_timepoint()} to register events, \code{get_timepoint()} for lookup,
\code{get_end_timepoint()} / \code{get_n_arms()} / \code{get_unique_times()} for
summary queries.
}
\details{
Trigger conditions (filtering + analysis) are now managed by the separate
\code{\link{Condition}} class. \code{Condition} objects are stored in \code{trial$conditions}
and evaluated by \code{\link{Trial}}\verb{$run()} at each timepoint.

Helper functions \code{\link[=condition_calendar_time]{condition_calendar_time()}} and \code{\link[=condition_enrollment_fraction]{condition_enrollment_fraction()}}
provide convenient shortcuts for building \code{Condition} objects; both return
a \code{\link{Condition}} that you pass to \code{Trial$new(conditions = list(...))}.
}
\examples{
# Basic construction
t <- Timer$new(name = "Timer")

# Add timepoints
t$add_timepoint(time = 1, arm = "A", drop = 2L, enroll = 10L)
t$add_timepoint(time = 2, arm = "A", drop = 1L, enroll = 12L)
t$add_timepoint(time = 1, arm = "B", drop = 0L, enroll = 8L)

# Query
t$get_end_timepoint() # max time => 2
t$get_n_arms()        # unique arms => 2
t$get_unique_times()  # unique times => c(1, 2)
t$get_timepoint("A", 1) # returns a single-row data.frame


## ------------------------------------------------
## Method `Timer$new`
## ------------------------------------------------

t <- Timer$new(name = "Timer")

## ------------------------------------------------
## Method `Timer$add_timepoint`
## ------------------------------------------------

t <- Timer$new(name = "Timer")
t$add_timepoint(time = 1, arm = "A", drop = 1L, enroll = 3L)

## ------------------------------------------------
## Method `Timer$get_end_timepoint`
## ------------------------------------------------

t <- Timer$new(name = "Timer")
t$add_timepoint(time = 3.14, arm = "A", drop = 7L, enroll = 22L)
t$get_end_timepoint()

## ------------------------------------------------
## Method `Timer$get_n_arms`
## ------------------------------------------------

t <- Timer$new(name = "Timer")
t$add_timepoint(time = 3.14, arm = "A", drop = 7L, enroll = 22L)
t$add_timepoint(time = 3.28, arm = "B", drop = 6L, enroll = 23L)
t$get_n_arms()

## ------------------------------------------------
## Method `Timer$get_unique_times`
## ------------------------------------------------

t <- Timer$new(name = "Timer")
t$add_timepoint(time = 3.14, arm = "A", drop = 7L, enroll = 22L)
t$add_timepoint(time = 3.28, arm = "B", drop = 6L, enroll = 23L)
t$get_unique_times()

## ------------------------------------------------
## Method `Timer$get_timepoint`
## ------------------------------------------------

t <- Timer$new(name = "Timer")
t$add_timepoint(time = 3.14, arm = "A", drop = 7L, enroll = 22L)
t$add_timepoint(time = 3.28, arm = "B", drop = 6L, enroll = 23L)
t$get_timepoint("A", 3.14)
}
\seealso{
\code{\link{Trial}} to coordinate simulations with populations,
\code{\link{Condition}} for trigger/analysis logic,
\code{\link[=add_timepoints]{add_timepoints()}} to attach multiple timepoints.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{name}}{\code{character} Unique identifier for the \code{Timer} instance.}

\item{\code{timelist}}{\code{data.frame} A data.frame of timepoints with columns:
\itemize{
\item \code{time} \code{numeric} Calendar time
\item \code{arm} \code{character} Unique identifier of the arm
\item \code{drop} \code{integer} # of subjects dropped at \code{time}
\item \code{enroll} \code{integer} # of subjects enrolled at \code{time}
}}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Timer-new}{\code{Timer$new()}}
\item \href{#method-Timer-add_timepoint}{\code{Timer$add_timepoint()}}
\item \href{#method-Timer-get_end_timepoint}{\code{Timer$get_end_timepoint()}}
\item \href{#method-Timer-get_n_arms}{\code{Timer$get_n_arms()}}
\item \href{#method-Timer-get_unique_times}{\code{Timer$get_unique_times()}}
\item \href{#method-Timer-get_timepoint}{\code{Timer$get_timepoint()}}
\item \href{#method-Timer-clone}{\code{Timer$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Timer-new"></a>}}
\if{latex}{\out{\hypertarget{method-Timer-new}{}}}
\subsection{Method \code{new()}}{
Create a new \code{Timer} instance.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Timer$new(name, timelist = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{\code{character} Unique identifier.}

\item{\code{timelist}}{\code{data.frame} Optional data.frame of timepoints with columns
\code{time}, \code{arm}, \code{drop}, \code{enroll}. If \code{NULL}, an empty frame is created.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{Timer} instance.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{t <- Timer$new(name = "Timer")
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Timer-add_timepoint"></a>}}
\if{latex}{\out{\hypertarget{method-Timer-add_timepoint}{}}}
\subsection{Method \code{add_timepoint()}}{
Add a timepoint to the timer.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Timer$add_timepoint(time, arm, drop = 0L, enroll)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{time}}{\code{numeric} Calendar time.}

\item{\code{arm}}{\code{character} Arm identifier.}

\item{\code{drop}}{\code{integer} Count of subjects to drop.}

\item{\code{enroll}}{\code{integer} Count of subjects to enroll.}
}
\if{html}{\out{</div>}}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{t <- Timer$new(name = "Timer")
t$add_timepoint(time = 1, arm = "A", drop = 1L, enroll = 3L)
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Timer-get_end_timepoint"></a>}}
\if{latex}{\out{\hypertarget{method-Timer-get_end_timepoint}{}}}
\subsection{Method \code{get_end_timepoint()}}{
Determine the last timepoint for a given instance of \code{Timer} class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Timer$get_end_timepoint()}\if{html}{\out{</div>}}
}

\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{t <- Timer$new(name = "Timer")
t$add_timepoint(time = 3.14, arm = "A", drop = 7L, enroll = 22L)
t$get_end_timepoint()
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Timer-get_n_arms"></a>}}
\if{latex}{\out{\hypertarget{method-Timer-get_n_arms}{}}}
\subsection{Method \code{get_n_arms()}}{
Get number of unique arms.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Timer$get_n_arms()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{integer} Number of unique arms.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{t <- Timer$new(name = "Timer")
t$add_timepoint(time = 3.14, arm = "A", drop = 7L, enroll = 22L)
t$add_timepoint(time = 3.28, arm = "B", drop = 6L, enroll = 23L)
t$get_n_arms()
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Timer-get_unique_times"></a>}}
\if{latex}{\out{\hypertarget{method-Timer-get_unique_times}{}}}
\subsection{Method \code{get_unique_times()}}{
Get unique timepoints.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Timer$get_unique_times()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{numeric} vector of unique times.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{t <- Timer$new(name = "Timer")
t$add_timepoint(time = 3.14, arm = "A", drop = 7L, enroll = 22L)
t$add_timepoint(time = 3.28, arm = "B", drop = 6L, enroll = 23L)
t$get_unique_times()
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Timer-get_timepoint"></a>}}
\if{latex}{\out{\hypertarget{method-Timer-get_timepoint}{}}}
\subsection{Method \code{get_timepoint()}}{
Get a timepoint by arm and time value.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Timer$get_timepoint(arm, i)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{arm}}{\code{character} Arm identifier.}

\item{\code{i}}{\code{numeric} Time value to look up.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Single-row \code{data.frame} or \code{NULL} if not found.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{t <- Timer$new(name = "Timer")
t$add_timepoint(time = 3.14, arm = "A", drop = 7L, enroll = 22L)
t$add_timepoint(time = 3.28, arm = "B", drop = 6L, enroll = 23L)
t$get_timepoint("A", 3.14)
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Timer-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Timer-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Timer$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
