% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{trigger_by_calendar}
\alias{trigger_by_calendar}
\title{Trigger Analysis at a Calendar Time}
\usage{
trigger_by_calendar(cal_time, timer, analysis = NULL)
}
\arguments{
\item{cal_time}{\code{numeric} Calendar time(s) to trigger.}

\item{timer}{\code{\link{Timer}} instance to update.}

\item{analysis}{\code{function} or \code{NULL} Optional function to apply.}
}
\value{
Invisible \code{\link{Timer}}.
}
\description{
Adds an analysis trigger at a specified calendar time.
}
\examples{
t <- Timer$new("timer")
trigger_by_calendar(2, t, analysis = function(df, current_time) nrow(df))

}
\seealso{
\link{Timer}, \code{\link[=trigger_by_fraction]{trigger_by_fraction()}}.
}
