% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{trigger_by_calendar}
\alias{trigger_by_calendar}
\title{Trigger Analysis at a Calendar Time}
\usage{
trigger_by_calendar(cal_time, analysis = NULL, name = NULL)
}
\arguments{
\item{cal_time}{\code{numeric} Calendar time(s) at which to trigger.}

\item{analysis}{\code{function} or \code{NULL} Optional analysis function called as
\code{analysis(filtered_data, current_time)}. If \code{NULL}, the filtered snapshot
is returned as-is with a warning.}

\item{name}{\code{character} or \code{NULL} Result key. Defaults to
\code{"cal_time_<cal_time>"}.}
}
\value{
A \code{\link{Condition}} object.
}
\description{
Builds a \code{\link{Condition}} that fires when the trial clock reaches a specified
calendar time. The returned \code{Condition} should be passed to
\code{Trial$new(conditions = list(...))}.
}
\examples{
cond <- trigger_by_calendar(
  cal_time = 12,
  analysis = function(df, current_time) {
    data.frame(n_enrolled = sum(!is.na(df$enroll_time)))
  }
)

}
\seealso{
\link{Condition}, \code{\link[=trigger_by_fraction]{trigger_by_fraction()}}, \link{Trial}.
}
