% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{trigger_by_fraction}
\alias{trigger_by_fraction}
\title{Trigger Analysis at a Sample Fraction}
\usage{
trigger_by_fraction(fraction, timer, sample_size, analysis = NULL)
}
\arguments{
\item{fraction}{\code{numeric} Sample fraction (0 < fraction <= 1).}

\item{timer}{\code{\link{Timer}} instance to update.}

\item{sample_size}{\code{integer} Target sample size.}

\item{analysis}{\code{function} or \code{NULL} Optional function to apply.}
}
\value{
Invisible \code{\link{Timer}}.
}
\description{
Adds an analysis trigger when a fraction of the target sample enrolls.
}
\examples{
t <- Timer$new("timer")
trigger_by_fraction(0.5, t, sample_size = 100, analysis = function(df, current_time) nrow(df))

}
\seealso{
\link{Timer}, \code{\link[=trigger_by_calendar]{trigger_by_calendar()}}.
}
