% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triggers.R
\name{trigger_primitives}
\alias{trigger_primitives}
\alias{value_trigger}
\alias{count_trigger}
\alias{enroll_trigger}
\alias{calendar_trigger}
\alias{&.rxsim_trigger}
\alias{|.rxsim_trigger}
\title{Build Safe Trial Triggers}
\usage{
value_trigger(col, op, rhs)

count_trigger(col, op, rhs)

enroll_trigger(fraction, sample_size)

calendar_trigger(cal_time)

\method{&}{rxsim_trigger}(e1, e2)

\method{|}{rxsim_trigger}(e1, e2)
}
\arguments{
\item{col}{\code{character} Column name referenced by the trigger.}

\item{op}{\code{character} Comparison operator. Must be one of
\code{c(">=", "<=", ">", "<", "==", "!=", "\%in\%")}.}

\item{rhs}{Right-hand side value. Must be atomic for \code{value_trigger()} and
numeric for \code{count_trigger()}.}

\item{fraction}{\code{numeric} Sample fraction (0 < fraction <= 1).}

\item{sample_size}{\code{numeric} Target sample size.}

\item{cal_time}{\code{numeric} Calendar time(s) at which to trigger.}

\item{e1, e2}{\code{rxsim_trigger} objects to combine.}
}
\value{
An \code{rxsim_trigger} object.
}
\description{
Create inert trigger specifications that can be safely
passed to \code{\link{Condition}} or composed with \code{&} and \code{|}.
}
\examples{
t1 <- value_trigger("time", ">=", 52)
t2 <- count_trigger("enroll_time", ">=", 100)
t3 <- enroll_trigger(0.5, 200) & calendar_trigger(52)
}
\seealso{
\link{Condition}, \code{enroll_trigger()}, \code{calendar_trigger()}.
}
