% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{vector_to_dataframe}
\alias{vector_to_dataframe}
\title{Convert Vector to Population Data Frame}
\usage{
vector_to_dataframe(data)
}
\arguments{
\item{data}{\code{numeric} vector of population values.}
}
\value{
\code{data.frame} with columns: \code{id}, \code{data}, \code{readout_time}.
}
\description{
Formats a numeric vector as a population data frame.
}
\seealso{
\link{Population}.
}
